♻️ refactor(toolregistry): reorganize websearch module structure
- move websearch-related classes into a dedicated subdirectory
- update module imports to reflect new structure
